Added docstrings in tseries.py
Docstrings are added in tseries.py. At the top of the page the minimum
requirements for each tseries class is defined.

The questions is whether we want to provide a elaborate docstring for
each tseries class (E.g. Tseries) or a short summary of each tseries
class (E.g. Tseries2).