Obs can now have different index as Tseries
When this happens, use linear interpolation to calculate the simulation
at the observation timesteps, when residuals need to be determined. Also
edited set_tmin_tmax, so that tmin and tmax are also determined by the
tseries, andd not just by the oseries. Now we also compare the
time-offset from the frequency between the tseries (now only works for
days). This value is saved in the new model-attribute dt. Lastly edited
the plot_decomposition, as sometimes the observations or grid-lines were
not shown.