Frequenty independent calculation
If everything went well, it is now possible to make models with
different frequencies, while the parameters stay roughly the same (so
the parameters allways have a daily unit).  The frequency is therefore
now passsed around between the different objects. The previously named
variable dt is now renamed to time_offset, to distingish it from the
variable dt that is calculated from the frequency in the
simulation-method.

Only the noise model variable 'delt' still needs to be updated, but I do
not understand what is does.