up keyword for tseries
Each tseries now has an ‘up’ keyword to indicate a positive stress
causes the head to go up. Default is True. Tseries2 is stress1 up and
stress2 down. I know Raoul thinks this is too restrictive, but that’s
what it is for now. Initial values of series are all set to give a step
of 1 for the mean stress.