Enhanced: TimeSeries
some optimization for when a TimeSeries object is provided as the series argument, taking the fastpath for creating the new TimeSeries. Mostly computational but this also solves a bug where the original data was not available anymore when creating a TimeSeries with a TimeSeries object..

TODO: update DataModel for the read lib.