noise model is automatically added
A NoiseModel is automatically added when a Model is instantiated. The
add_model function still works at it replaces the model that is added
on instantiation, but is not needed anymore. All Notebooks are changed
accordingly, but not the regular Python examples. Solving without a
NoiseModel still works when specifying `ml.solve(noise=False)`