added noisemodel as a keyword in model
Added noisemodel as a keyword.
This could in the future be built towards a keyword that is either boolean or takes a noisemodel instance.

Also, constant can still be None, as fixing this parameters is not possible with all solver so it should be possible to not have a constant.