new sample up method: "divide"
New method to sample up time series (e.g. vweek to day). The values are divided over period. This method is standard for the well time series which are often provided in intensities (e.g. m3/month or m3).  Sample down is now sum for wells. This ensures the parameter values from the response functions can be interpreted easily.