new method get odelt and speed increase
introduced the method get_odelt. This makes it possible (in the future) to change the oseries (e.g. freq) when optimizing. Also added some improvements in speed by adding pandas methods and removing redundant code. Combined with the noise model this reduces the time to solve by 30% for the example.py