Keep oseries_calib and sim_index
These two model-attributes can be used for plotting, and are only recalculated when needed. Improved the obeservations() method (which is used in the pliots and stats module), so that it returns the observations that are used in calibration. The observations that are used by the solver are plotted in black, and those that are not used by the solver are plotted in grey.