Restuctured code in Model class
Restructured a large part of the the Model class and other files to be more readable and easier to understand what is going in. This involved removing redundant code (such as removing get_oseries_calib which could be combined with the observations method), renaming variables to be more consistent and making some names easier to understand. Basically, nothing changed (except internally), except that I also addressed issue #70 , which is now solved.

- updated get_sim_index
- removed get_oseries_calib
- simulation is called sim everywhere (renamed h => sim in simulate), residuals are res, noise is noise. odelt is odelt. stressmodels as sm. Still need to do oseries and oseries_calib...
- ml.simulate respects tmin. extra keyword is added to obtain the simulation in the warmup period.