Improved Kraijenhoff van de Leur response function
Improved the Kraijenhoff van de Leur response function. The function now implemented is actually Bruggeman equation 133.15. This function has a gain which can be calculated and is a bit faster. Also Numba is not needed.

Also adjusted the name from Kleur to Kraijenhoff (which will cause a lot of typos but is more accurate).